Trying some cleanup refactoring, may be reverted based on review.
I'm unsure if I want the Operator constructor to take all of request, db params, and accession code as parameters or if I want the methods to take some or all of these values as parameters. Putting what I have up here to start and will change based on feedback.
