+++ vegbank/operators/TaxonObservationOperator.py
@@ -1,0 +1,57 @@
+from flask import jsonify, request, send_file
+import psycopg
+from psycopg import connect, ClientCursor
+from psycopg.rows import dict_row
+import pandas as pd
+import numpy as np
+import io
+import time
+import operators.table_defs_config as table_defs_config
+import traceback
+from operators.operator_parent_class import Operator
+from utilities import jsonify_error_message, convert_to_parquet, allowed_file
+
+class TaxonObservationOperator(Operator):
+    def __init__(self, request, params, accession_code):
+        super().__init__(request, params, accession_code)
+
+    def get_taxon_observations(self, accession_code):
+        detail = request.args.get("detail", self.default_detail)
+        if detail not in ("minimal", "full"):
+            return jsonify_error_message("When provided, 'detail' must be 'minimal' or 'full'."), 400
+        try:
+            limit = int(request.args.get("limit", self.default_limit))
+            offset = int(request.args.get("offset", self.default_offset))
+            num_taxa = int(request.args.get("num_taxa", 5))  # Default to 5 if not specified
+        except ValueError:
+            return jsonify_error_message("When provided, 'offset' 'numTaxa' and 'limit' must be non-negative integers."), 400
+        
+        data = (num_taxa, limit, offset, )
+        
+        with open(self.QUERIES_FOLDER + "/taxon_observation/get_top_taxa_count.sql", "r") as file:
+            count_sql = file.read()
+        countData = (num_taxa, )
+
+        sql = ""
+        if(accession_code is None):
+            with open(self.QUERIES_FOLDER + "/taxon_observation/get_top_taxa_coverage.sql", "r") as file:
+                sql = file.read()
+        else: #TODO This either needs to be an observation accession code, or a taxa one.
+            with open(self.QUERIES_FOLDER + "/taxon_observation/get_taxa_by_accession_code.sql", "r") as file:
+                sql = file.read()
+                data = (accession_code, )
+
+        to_return = {}
+        with psycopg.connect(**self.params, row_factory=dict_row) as conn:
+            with conn.cursor() as cur:
+                cur.execute(sql, data)
+                to_return["data"] = cur.fetchall()
+                print("number of records")
+
+                #if(accession_code is None):
+                #    cur.execute(count_sql, countData)
+                #    to_return["count"] = cur.fetchall()[0]["count"]
+                #else:
+                #    to_return["count"] = len(to_return["data"])
+            conn.close()      
+        return jsonify(to_return)
